requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
phidata==2.7.6
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
youtube-transcript-api==0.6.3
